+++ service/src/main/java/com/topjohnwu/superuser/internal/RootServiceManager.java
@@ -21,6 +21,7 @@ import static com.topjohnwu.superuser.internal.RootServerMain.CMDLINE_START_SERV
 import static com.topjohnwu.superuser.internal.RootServerMain.CMDLINE_STOP_SERVICE;
 import static com.topjohnwu.superuser.ipc.RootService.CATEGORY_DAEMON_MODE;
 
+import android.Manifest;
 import android.annotation.SuppressLint;
 import android.content.BroadcastReceiver;
 import android.content.ComponentName;
@@ -58,9 +59,13 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import java.util.UUID;
 import java.util.concurrent.Executor;
 
+/**
+ * Runs in the non-root (client) process.
+ *
+ * Starts the root process and manages connections with the remote process.
+ */
 @RestrictTo(RestrictTo.Scope.LIBRARY)
 public class RootServiceManager implements Handler.Callback {
 
@@ -75,6 +80,7 @@ public class RootServiceManager implements Handler.Callback {
     private static final String BUNDLE_BINDER_KEY = "binder";
     private static final String INTENT_BUNDLE_KEY = "extra.bundle";
     private static final String INTENT_DAEMON_KEY = "extra.daemon";
+    private static final String RECEIVER_BROADCAST = "com.topjohnwu.superuser.RECEIVER_BROADCAST";
     private static final String API_27_DEBUG =
             "-Xrunjdwp:transport=dt_android_adb,suspend=n,server=y " +
             "-Xcompiler-option --debuggable";
@@ -91,6 +97,7 @@ public class RootServiceManager implements Handler.Callback {
 
     private static final int REMOTE_EN_ROUTE = 1 << 0;
     private static final int DAEMON_EN_ROUTE = 1 << 1;
+    private static final int RECEIVER_REGISTERED = 1 << 2;
 
     public static RootServiceManager getInstance() {
         if (mInstance == null) {
@@ -103,7 +110,7 @@ public class RootServiceManager implements Handler.Callback {
     static Intent getBroadcastIntent(IBinder binder, boolean isDaemon) {
         Bundle bundle = new Bundle();
         bundle.putBinder(BUNDLE_BINDER_KEY, binder);
-        return new Intent()
+        return new Intent(RECEIVER_BROADCAST)
                 .setPackage(Utils.context.getPackageName())
                 .addFlags(HiddenAPIs.FLAG_RECEIVER_FROM_SHELL)
                 .putExtra(INTENT_DAEMON_KEY, isDaemon)
@@ -117,7 +124,7 @@ public class RootServiceManager implements Handler.Callback {
     }
 
     @NonNull
-    private static Pair<ComponentName, Boolean> enforceIntent(Intent intent) {
+    private static ServiceKey parseIntent(Intent intent) {
         ComponentName name = intent.getComponent();
         if (name == null) {
             throw new IllegalArgumentException("The intent does not have a component set");
@@ -125,28 +132,21 @@ public class RootServiceManager implements Handler.Callback {
         if (!name.getPackageName().equals(Utils.getContext().getPackageName())) {
             throw new IllegalArgumentException("RootServices outside of the app are not supported");
         }
-        return new Pair<>(name, intent.hasCategory(CATEGORY_DAEMON_MODE));
-    }
-
-    private static void notifyDisconnection(
-            Map.Entry<ServiceConnection, Pair<RemoteService, Executor>> e) {
-        ServiceConnection c = e.getKey();
-        ComponentName name = e.getValue().first.key.first;
-        e.getValue().second.execute(() -> c.onServiceDisconnected(name));
+        return new ServiceKey(name, intent.hasCategory(CATEGORY_DAEMON_MODE));
     }
 
     private RemoteProcess mRemote;
     private RemoteProcess mDaemon;
 
-    private String filterAction;
     private int flags = 0;
 
     private final List<BindTask> pendingTasks = new ArrayList<>();
-    private final Map<Pair<ComponentName, Boolean>, RemoteService> services = new ArrayMap<>();
-    private final Map<ServiceConnection, Pair<RemoteService, Executor>> connections = new ArrayMap<>();
+    private final Map<ServiceKey, RemoteServiceRecord> services = new ArrayMap<>();
+    private final Map<ServiceConnection, ConnectionRecord> connections = new ArrayMap<>();
 
     private RootServiceManager() {}
 
+    @SuppressLint("InlinedApi")
     private Shell.Task startRootProcess(ComponentName name, String action) {
         Context context = Utils.getContext();
 
@@ -154,11 +154,21 @@ public class RootServiceManager implements Handler.Callback {
             Log.e(TAG, JVMTI_ERROR);
         }
 
-        if (filterAction == null) {
-            filterAction = UUID.randomUUID().toString();
+        if ((flags & RECEIVER_REGISTERED) == 0) {
             // Register receiver to receive binder from root process
-            IntentFilter filter = new IntentFilter(filterAction);
-            context.registerReceiver(new ServiceReceiver(), filter);
+            IntentFilter filter = new IntentFilter(RECEIVER_BROADCAST);
+            // Guard the receiver behind permission BROADCAST_PACKAGE_REMOVED. This permission
+            // is not obtainable by normal apps, making the receiver effectively non-exported.
+            // On Android 13+, we can also rely on the flag RECEIVER_NOT_EXPORTED.
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                context.registerReceiver(new ServiceReceiver(), filter,
+                        Manifest.permission.BROADCAST_PACKAGE_REMOVED, null,
+                        Context.RECEIVER_NOT_EXPORTED);
+            } else {
+                context.registerReceiver(new ServiceReceiver(), filter,
+                        Manifest.permission.BROADCAST_PACKAGE_REMOVED, null);
+            }
+            flags |= RECEIVER_REGISTERED;
         }
 
         return (stdin, stdout, stderr) -> {
@@ -191,15 +201,28 @@ public class RootServiceManager implements Handler.Callback {
                 }
             }
 
+            final String niceNameCmd;
+            switch (action) {
+                case CMDLINE_START_SERVICE:
+                    niceNameCmd = String.format(Locale.ROOT, "--nice-name=%s:root:%d",
+                            ctx.getPackageName(), Process.myUid() / 100000);
+                    break;
+                case CMDLINE_START_DAEMON:
+                    niceNameCmd = "--nice-name=" + ctx.getPackageName() + ":root:daemon";
+                    break;
+                default:
+                    niceNameCmd = "";
+                    break;
+            }
+
             String app_process = new File("/proc/self/exe").getCanonicalPath();
             String cmd = String.format(Locale.ROOT,
-                    "(%s CLASSPATH=%s %s %s /system/bin --nice-name=%s:root " +
-                    "com.topjohnwu.superuser.internal.RootServerMain %s %d %s %s >/dev/null 2>&1)&",
-                    env, mainJar, app_process, params, ctx.getPackageName(),
-                    name.flattenToString().replace("$", "\\$"), // args[0]
-                    Process.myUid(),                            // args[1]
-                    filterAction,                               // args[2]
-                    action);                                    // args[3]
+                    "(%s CLASSPATH=%s %s %s /system/bin %s " +
+                    "com.topjohnwu.superuser.internal.RootServerMain '%s' %d %s >/dev/null 2>&1)&",
+                    env, mainJar, app_process, params, niceNameCmd,
+                    name.flattenToString(),   // args[0]
+                    Process.myUid(),          // args[1]
+                    action);                  // args[2]
 
             Utils.log(TAG, cmd);
             // Write command to stdin
@@ -214,33 +237,33 @@ public class RootServiceManager implements Handler.Callback {
     }
 
     // Returns null if binding is done synchronously, or else return key
-    private Pair<ComponentName, Boolean> bindInternal(
-            Intent intent, Executor executor, ServiceConnection conn) {
+    private ServiceKey bindInternal(Intent intent, Executor executor, ServiceConnection conn) {
         enforceMainThread();
 
         // Local cache
-        Pair<ComponentName, Boolean> key = enforceIntent(intent);
-        RemoteService s = services.get(key);
+        ServiceKey key = parseIntent(intent);
+        RemoteServiceRecord s = services.get(key);
         if (s != null) {
-            connections.put(conn, new Pair<>(s, executor));
+            connections.put(conn, new ConnectionRecord(s, executor));
             s.refCount++;
-            executor.execute(() -> conn.onServiceConnected(key.first, s.binder));
+            IBinder binder = s.binder;
+            executor.execute(() -> conn.onServiceConnected(key.getName(), binder));
             return null;
         }
 
-        RemoteProcess p = key.second ? mDaemon : mRemote;
+        RemoteProcess p = key.isDaemon() ? mDaemon : mRemote;
         if (p == null)
             return key;
 
         try {
-            IBinder binder = p.sm.bind(intent);
+            IBinder binder = p.mgr.bind(intent);
             if (binder != null) {
-                RemoteService r = new RemoteService(key, binder, p);
-                connections.put(conn, new Pair<>(r, executor));
-                services.put(key, r);
-                executor.execute(() -> conn.onServiceConnected(key.first, binder));
+                s = new RemoteServiceRecord(key, binder, p);
+                connections.put(conn, new ConnectionRecord(s, executor));
+                services.put(key, s);
+                executor.execute(() -> conn.onServiceConnected(key.getName(), binder));
             } else if (Build.VERSION.SDK_INT >= 28) {
-                executor.execute(() -> conn.onNullBinding(key.first));
+                executor.execute(() -> conn.onNullBinding(key.getName()));
             }
         } catch (RemoteException e) {
             Utils.err(TAG, e);
@@ -252,14 +275,14 @@ public class RootServiceManager implements Handler.Callback {
     }
 
     public Shell.Task createBindTask(Intent intent, Executor executor, ServiceConnection conn) {
-        Pair<ComponentName, Boolean> key = bindInternal(intent, executor, conn);
+        ServiceKey key = bindInternal(intent, executor, conn);
         if (key != null) {
             pendingTasks.add(() -> bindInternal(intent, executor, conn) == null);
-            int mask = key.second ? DAEMON_EN_ROUTE : REMOTE_EN_ROUTE;
-            String action = key.second ? CMDLINE_START_DAEMON : CMDLINE_START_SERVICE;
+            int mask = key.isDaemon() ? DAEMON_EN_ROUTE : REMOTE_EN_ROUTE;
             if ((flags & mask) == 0) {
                 flags |= mask;
-                return startRootProcess(key.first, action);
+                String action = key.isDaemon() ? CMDLINE_START_DAEMON : CMDLINE_START_SERVICE;
+                return startRootProcess(key.getName(), action);
             }
         }
         return null;
@@ -268,76 +291,98 @@ public class RootServiceManager implements Handler.Callback {
     public void unbind(@NonNull ServiceConnection conn) {
         enforceMainThread();
 
-        Pair<RemoteService, Executor> p = connections.remove(conn);
-        if (p != null) {
-            p.first.refCount--;
-            p.second.execute(() -> conn.onServiceDisconnected(p.first.key.first));
-            if (p.first.refCount == 0) {
+        ConnectionRecord r = connections.remove(conn);
+        if (r != null) {
+            RemoteServiceRecord s = r.getService();
+            s.refCount--;
+            if (s.refCount == 0) {
                 // Actually close the service
-                services.remove(p.first.key);
+                services.remove(s.key);
                 try {
-                    p.first.host.sm.unbind(p.first.key.first);
+                    s.host.mgr.unbind(s.key.getName());
                 } catch (RemoteException e) {
                     Utils.err(TAG, e);
                 }
             }
+            r.disconnect(conn);
         }
     }
 
-    private void stopInternal(Pair<ComponentName, Boolean> key) {
-        RemoteService s = services.remove(key);
-        if (s == null)
-            return;
-
-        // Notify all connections
-        Iterator<Map.Entry<ServiceConnection, Pair<RemoteService, Executor>>> it =
+    private void dropConnections(Predicate predicate) {
+        Iterator<Map.Entry<ServiceConnection, ConnectionRecord>> it =
                 connections.entrySet().iterator();
         while (it.hasNext()) {
-            Map.Entry<ServiceConnection, Pair<RemoteService, Executor>> e = it.next();
-            if (e.getValue().first.equals(s)) {
-                notifyDisconnection(e);
+            Map.Entry<ServiceConnection, ConnectionRecord> e = it.next();
+            ConnectionRecord r = e.getValue();
+            if (predicate.eval(r.getService())) {
+                r.disconnect(e.getKey());
                 it.remove();
             }
         }
     }
 
+    private void onServiceStopped(ServiceKey key) {
+        RemoteServiceRecord s = services.remove(key);
+        if (s != null)
+            dropConnections(s::equals);
+    }
+
     public Shell.Task createStopTask(Intent intent) {
         enforceMainThread();
 
-        Pair<ComponentName, Boolean> key = enforceIntent(intent);
-        RemoteProcess p = key.second ? mDaemon : mRemote;
+        ServiceKey key = parseIntent(intent);
+        RemoteProcess p = key.isDaemon() ? mDaemon : mRemote;
         if (p == null) {
-            if (key.second) {
+            if (key.isDaemon()) {
                 // Start a new root process to stop daemon
-                return startRootProcess(key.first, CMDLINE_STOP_SERVICE);
+                return startRootProcess(key.getName(), CMDLINE_STOP_SERVICE);
             }
             return null;
         }
 
-        stopInternal(key);
         try {
-            p.sm.stop(key.first, -1, null);
+            p.mgr.stop(key.getName(), -1);
         } catch (RemoteException e) {
             Utils.err(TAG, e);
         }
+
+        onServiceStopped(key);
         return null;
     }
 
     @Override
     public boolean handleMessage(@NonNull Message msg) {
         if (msg.what == MSG_STOP) {
-            stopInternal(new Pair<>((ComponentName) msg.obj, msg.arg1 != 0));
+            onServiceStopped(new ServiceKey((ComponentName) msg.obj, msg.arg1 != 0));
         }
         return false;
     }
 
-    class RemoteProcess extends BinderHolder {
+    private static class ServiceKey extends Pair<ComponentName, Boolean> {
+        ServiceKey(ComponentName name, boolean isDaemon) {
+            super(name, isDaemon);
+        }
+        ComponentName getName() { return first; }
+        boolean isDaemon() { return second; }
+    }
 
-        final IRootServiceManager sm;
+    private static class ConnectionRecord extends Pair<RemoteServiceRecord, Executor> {
+        ConnectionRecord(RemoteServiceRecord s, Executor e) {
+            super(s, e);
+        }
+        RemoteServiceRecord getService() { return first; }
+        void disconnect(ServiceConnection conn) {
+            second.execute(() -> conn.onServiceDisconnected(first.key.getName()));
+        }
+    }
+
+    private class RemoteProcess extends BinderHolder {
+
+        final IRootServiceManager mgr;
 
         RemoteProcess(IRootServiceManager s) throws RemoteException {
             super(s.asBinder());
-            sm = s;
+            mgr = s;
         }
 
         @Override
@@ -347,26 +392,30 @@ public class RootServiceManager implements Handler.Callback {
             if (mDaemon == this)
                 mDaemon = null;
 
-            Iterator<RemoteService> sit = services.values().iterator();
+            Iterator<RemoteServiceRecord> sit = services.values().iterator();
             while (sit.hasNext()) {
                 if (sit.next().host == this) {
                     sit.remove();
                 }
             }
-
-            Iterator<Map.Entry<ServiceConnection, Pair<RemoteService, Executor>>> it =
-                    connections.entrySet().iterator();
-            while (it.hasNext()) {
-                Map.Entry<ServiceConnection, Pair<RemoteService, Executor>> e = it.next();
-                if (e.getValue().first.host == this) {
-                    notifyDisconnection(e);
-                    it.remove();
-                }
-            }
+            dropConnections(s -> s.host == this);
         }
     }
 
-    class ServiceReceiver extends BroadcastReceiver {
+    private static class RemoteServiceRecord {
+        final ServiceKey key;
+        final IBinder binder;
+        final RemoteProcess host;
+        int refCount = 1;
+
+        RemoteServiceRecord(ServiceKey key, IBinder binder, RemoteProcess host) {
+            this.key = key;
+            this.binder = binder;
+            this.host = host;
+        }
+    }
+
+    private class ServiceReceiver extends BroadcastReceiver {
 
         private final Messenger m;
 
@@ -385,10 +434,10 @@ public class RootServiceManager implements Handler.Callback {
             if (binder == null)
                 return;
 
-            IRootServiceManager sm = IRootServiceManager.Stub.asInterface(binder);
+            IRootServiceManager mgr = IRootServiceManager.Stub.asInterface(binder);
             try {
-                sm.connect(m.getBinder());
-                RemoteProcess p = new RemoteProcess(sm);
+                mgr.connect(m.getBinder());
+                RemoteProcess p = new RemoteProcess(mgr);
                 if (intent.getBooleanExtra(INTENT_DAEMON_KEY, false)) {
                     mDaemon = p;
                     flags &= ~DAEMON_EN_ROUTE;
@@ -407,20 +456,11 @@ public class RootServiceManager implements Handler.Callback {
         }
     }
 
-    static class RemoteService {
-        final Pair<ComponentName, Boolean> key;
-        final IBinder binder;
-        final RemoteProcess host;
-        int refCount = 1;
-
-        RemoteService(Pair<ComponentName, Boolean> key, IBinder binder, RemoteProcess host) {
-            this.key = key;
-            this.binder = binder;
-            this.host = host;
-        }
+    private interface BindTask {
+        boolean run();
     }
 
-    interface BindTask {
-        boolean run();
+    private interface Predicate {
+        boolean eval(RemoteServiceRecord s);
     }
 }